Rename parameter panel file
`_widget_parameter_panel.py` to `_widget_parameters_panel.py`
