+++ src/scwidgets/code/__init__.py
@@ -1,5 +1,5 @@
 from ._widget_code_input import CodeInput
-from ._widget_parameter_panel import ParametersPanel
+from ._widget_parameters_panel import ParametersPanel
 
 __all__ = [
     "CodeInput",
