+++ src/scwidgets/code/_widget_parameters_panel.py
